api : split&refine sdk and test
